refactor: change api code into api file (#16)
