+++ src/types/wish.ts
@@ -1,3 +1,4 @@
+import type { LocalTime } from "./time";
 
 export interface WishItem {
   wishId: number;
@@ -10,16 +11,11 @@ export interface WishItem {
   claimTotalCount: number;
 }
 
-import axiosInstance from "@/apis/axiosInstance";
-import type { LocalTime } from "./time";
-
 export type WishUpsertRequest = {
   title: string;
   itemName: string;
   brand?: string;
-  //
   quantity: number; // -> 이거 어떻게 할지.... 나중에 수정할 듯??/? --> 고기
-  //
   description?: string;
   openTime: string;
   closeTime: string;
@@ -47,12 +43,3 @@ export type WishCreateAndMatchResponse = {
     matches: WishMatchItem[];
   };
 };
-
-export const postWishMatch = (
-  payload: WishUpsertRequest,
-  params?: { radiusKm?: number; size?: number }
-) =>
-  axiosInstance.post<WishCreateAndMatchResponse>("/wish/match", payload, {
-    params: { radiusKm: params?.radiusKm ?? 3, size: params?.size ?? 10 },
-  });
-
